further mugene2 fixes to get mars*.mugene compile. (Codex)
----

I changed the front end in src/mugene2/detail/compiler_internal.cpp (line 1) and src/mugene2/detail/compiler_internal.hpp (line 1) so that:

- #define aliases are applied before track tokenization, which makes %STEP_DELTA_A and similar step-length substitutions compile correctly.
- macro target lists reuse the normal range parser, so definitions like #macro 1-19,24-31 CH ... are accepted.
- negative lead-in ticks from constructs like r%-1 no longer break output generation; SMF output is normalized, and SMF2 clip output preserves the lead-in as clip position.
- clip positions are now signed in include/mugene2/mugene2.hpp (line 1), include/augene2/project.hpp (line 1), and the save/load path in src/augene2/storage.cpp (line 1).

+++ src/augene2/storage.cpp
@@ -167,7 +167,7 @@ choc::value::Value saveProjectToJson(const Project& project) {
             ++clip_ordinal;
             auto clip_json = choc::value::createObject("AugeneClip");
             clip_json.addMember("kind", clipKindToString(clip.kind));
-            clip_json.addMember("position_dctpq", static_cast<int64_t>(clip.position_dctpq));
+            clip_json.addMember("position_dctpq", clip.position_dctpq);
             clip_json.addMember("name", clip.name);
 
             std::string clip_file = clip.file;
@@ -313,7 +313,7 @@ std::unique_ptr<Project> UapmdProjectStorage::load(const std::filesystem::path& 
                 ProjectClip clip;
                 clip.kind = clipKindFromJson(clip_json);
                 if (clip_json.hasObjectMember("position_dctpq"))
-                    clip.position_dctpq = static_cast<uint64_t>(clip_json["position_dctpq"].getWithDefault<int64_t>(0));
+                    clip.position_dctpq = clip_json["position_dctpq"].getWithDefault<int64_t>(0);
                 if (clip_json.hasObjectMember("name"))
                     clip.name = std::string(clip_json["name"].getString());
                 if (clip_json.hasObjectMember("file"))
